refactor(#233): dissolve TeammateReadModel into read-helpers free functions
The TeammateReadModel class held no mutable state — it was pure
input->output read projection wrapped in a class. Per the v3 refactor plan
(R1), dissolve it into free functions and move the read orchestration onto
the collection that owns the stores.

- New teammate-collection/read-helpers.ts with pure free functions:
  toStatus / toRecordRow (projection), foldLastTurns (returns the turns
  array only; the caller validates requestedTurns and assembles the
  TeamMateLastResult), assertInRoster / identityInRoster (scope
  adjudication, now taking dispatcherId as a param), plus the relocated
  module helpers (matchesRecordQuery / cursor codec / clampHistoryLimit /
  previewText / validateLastTurns).
- TeammateCollection owns list / status / history / last / mustIdentity /
  rosterList as methods, resolving the live runtime via
  this.entities.get(name)?.getRuntime() (dropping the runtimeFor closure).
- TeammateService uses the free functions for status() / last() / close()
  and the ensureStarted roster guard; the redundant teamId param on
  last()/toStatus is gone (it reads identity.team_id).
- createDispatcherAgent no longer constructs a read model; its entity uses
  the same free functions.
- readModel removed from TeammateServiceDeps; read-model.ts deleted.

Both collection.last(name) and entity.last() are kept (different
signatures and responsibilities); both fold via foldLastTurns + toStatus.
